refactor(GCS+gRPC): split `DefaultOptionsGrpc()` (#14090)
