+++ google/cloud/storage/google_cloud_cpp_storage_grpc.bzl
@@ -63,6 +63,7 @@ google_cloud_cpp_storage_grpc_hdrs = [
     "internal/grpc/channel_refresh.h",
     "internal/grpc/configure_client_context.h",
     "internal/grpc/ctype_cord_workaround.h",
+    "internal/grpc/default_options.h",
     "internal/grpc/hmac_key_metadata_parser.h",
     "internal/grpc/hmac_key_request_parser.h",
     "internal/grpc/make_cord.h",
@@ -122,6 +123,7 @@ google_cloud_cpp_storage_grpc_srcs = [
     "internal/grpc/buffer_read_object_data.cc",
     "internal/grpc/channel_refresh.cc",
     "internal/grpc/configure_client_context.cc",
+    "internal/grpc/default_options.cc",
     "internal/grpc/hmac_key_metadata_parser.cc",
     "internal/grpc/hmac_key_request_parser.cc",
     "internal/grpc/make_cord.cc",
