+++ google/cloud/storage/internal/async/default_options.cc
@@ -17,6 +17,7 @@
 #include "google/cloud/storage/async/options.h"
 #include "google/cloud/storage/async/resume_policy.h"
 #include "google/cloud/storage/async/writer_connection.h"
+#include "google/cloud/storage/internal/grpc/default_options.h"
 #include "google/cloud/storage/internal/grpc/stub.h"
 #include <limits>
