+++ google/cloud/storage/internal/grpc/stub_test.cc
@@ -14,16 +14,12 @@
 
 #include "google/cloud/storage/internal/grpc/stub.h"
 #include "google/cloud/storage/grpc_plugin.h"
-#include "google/cloud/storage/options.h"
 #include "google/cloud/storage/testing/mock_storage_stub.h"
 #include "google/cloud/credentials.h"
-#include "google/cloud/grpc_options.h"
 #include "google/cloud/options.h"
 #include "google/cloud/testing_util/is_proto_equal.h"
-#include "google/cloud/testing_util/scoped_environment.h"
 #include "google/cloud/testing_util/status_matchers.h"
 #include "google/cloud/testing_util/validate_metadata.h"
-#include "google/cloud/universe_domain_options.h"
 #include <google/protobuf/text_format.h>
 #include <gmock/gmock.h>
 
@@ -41,7 +37,6 @@ using ::google::cloud::storage::QuotaUser;
 using ::google::cloud::storage::internal::CreateNullHashFunction;
 using ::google::cloud::storage::testing::MockInsertStream;
 using ::google::cloud::storage::testing::MockStorageStub;
-using ::google::cloud::testing_util::ScopedEnvironment;
 using ::google::cloud::testing_util::StatusIs;
 using ::google::cloud::testing_util::ValidateMetadataFixture;
 using ::testing::Contains;
@@ -80,92 +75,6 @@ std::unique_ptr<GrpcStub> CreateTestClient(
   std::shared_ptr<google::cloud::internal::MinimalIamCredentialsStub> unused;
   return std::make_unique<GrpcStub>(std::move(stub), /*iam=*/unused,
                                     TestOptions());
-}
-
-TEST(DefaultOptionsGrpc, DefaultOptionsGrpcChannelCount) {
-  using ::google::cloud::GrpcNumChannelsOption;
-  struct TestCase {
-    std::string endpoint;
-    int lower_bound;
-    int upper_bound;
-  } cases[] = {
-      {"storage.googleapis.com", 4, std::numeric_limits<int>::max()},
-      {"google-c2p:///storage.googleapis.com", 1, 1},
-      {"google-c2p-experimental:///storage.googleapis.com", 1, 1},
-  };
-
-  for (auto const& test : cases) {
-    SCOPED_TRACE("Testing with " + test.endpoint);
-    auto opts =
-        DefaultOptionsGrpc(TestOptions().set<EndpointOption>(test.endpoint));
-    auto const count = opts.get<GrpcNumChannelsOption>();
-    EXPECT_LE(test.lower_bound, count);
-    EXPECT_GE(test.upper_bound, count);
-
-    auto override = DefaultOptionsGrpc(TestOptions()
-                                           .set<EndpointOption>(test.endpoint)
-                                           .set<GrpcNumChannelsOption>(42));
-    EXPECT_EQ(42, override.get<GrpcNumChannelsOption>());
-  }
-}
-
-TEST(DefaultOptionsGrpc, DefaultEndpoints) {
-  auto options = DefaultOptionsGrpc();
-  EXPECT_EQ(options.get<EndpointOption>(), "storage.googleapis.com");
-  EXPECT_EQ(options.get<AuthorityOption>(), "storage.googleapis.com");
-}
-
-TEST(DefaultOptionsGrpc, EndpointOptionsOverrideDefaults) {
-  ScopedEnvironment ud("GOOGLE_CLOUD_UNIVERSE_DOMAIN", "ud-env-var.net");
-
-  auto options = DefaultOptionsGrpc(
-      Options{}
-          .set<EndpointOption>("from-option")
-          .set<AuthorityOption>("host-from-option")
-          .set<internal::UniverseDomainOption>("ud-option.net"));
-  EXPECT_EQ(options.get<EndpointOption>(), "from-option");
-  EXPECT_EQ(options.get<AuthorityOption>(), "host-from-option");
-}
-
-TEST(DefaultOptionsGrpc, EnvVarsOverrideOptionsAndDefaults) {
-  ScopedEnvironment e("CLOUD_STORAGE_EXPERIMENTAL_GRPC_TESTBENCH_ENDPOINT",
-                      "from-testbench-env");
-  ScopedEnvironment ud("GOOGLE_CLOUD_UNIVERSE_DOMAIN", "ud-env-var.net");
-
-  auto options = DefaultOptionsGrpc(
-      Options{}
-          .set<EndpointOption>("from-ep-option")
-          .set<internal::UniverseDomainOption>("ud-option.net"));
-  EXPECT_EQ(options.get<EndpointOption>(), "from-testbench-env");
-  EXPECT_EQ(options.get<AuthorityOption>(), "storage.ud-env-var.net");
-}
-
-TEST(DefaultOptionsGrpc, IncorporatesUniverseDomain) {
-  auto options = DefaultOptionsGrpc(
-      Options{}.set<internal::UniverseDomainOption>("my-ud.net"));
-  EXPECT_EQ(options.get<EndpointOption>(), "storage.my-ud.net");
-  EXPECT_EQ(options.get<AuthorityOption>(), "storage.my-ud.net");
-}
-
-TEST(DefaultOptionsGrpc, IncorporatesUniverseDomainEnvVar) {
-  ScopedEnvironment ud("GOOGLE_CLOUD_UNIVERSE_DOMAIN", "ud-env-var.net");
-
-  auto options = DefaultOptionsGrpc(
-      Options{}.set<internal::UniverseDomainOption>("ud-option.net"));
-  EXPECT_EQ(options.get<EndpointOption>(), "storage.ud-env-var.net");
-  EXPECT_EQ(options.get<AuthorityOption>(), "storage.ud-env-var.net");
-}
-
-TEST(DefaultOptionsGrpc, DefaultOptionsUploadBuffer) {
-  auto const with_defaults =
-      DefaultOptionsGrpc(Options{}).get<storage::UploadBufferSizeOption>();
-  EXPECT_GE(with_defaults, 32 * 1024 * 1024L);
-
-  auto const with_override =
-      DefaultOptionsGrpc(
-          Options{}.set<storage::UploadBufferSizeOption>(256 * 1024))
-          .get<storage::UploadBufferSizeOption>();
-  EXPECT_EQ(with_override, 256 * 1024L);
 }
 
 TEST_F(GrpcClientTest, QueryResumableUpload) {
